// app/client/reducers/auth/types.ts
export enum LoginActionTypes {
  REAUTH_REQUEST = "REAUTH_REQUEST",
  LOGIN_REQUEST = "LOGIN_REQUEST",
  UPDATE_ACTIVE_ORG = "UPDATE_ACTIVE_ORG",
  LOGOUT = "LOGOUT",
  API_LOGOUT = "API_LOGOUT",
  LOGIN_PARTIAL = "LOGIN_PARTIAL",
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  LOGIN_FAILURE = "LOGIN_FAILURE",
}

export interface UpdateActiveOrgPayload {
  organisationIds: number[];
  isManageWallet?: boolean;
}

export interface LoginRequestPayload {
  body: {
    username: string;
    password: string;
  };
}
// ...
export interface AdminUser {
  id: number;
  email: string;
  admin_tier: string;
  created: string;
  is_activated: boolean;
  is_disabled: boolean;
}

export interface AdminUserByIDs {
  [key: number]: AdminUser;
}

export const UPDATE_ADMIN_USER_LIST = "UPDATE_ADMIN_USER_LIST";

export interface Invite {
  id: number;
  email: string;
  admin_tier: string;
  created: string;
}
// ...
export interface TokenData {
  auth_token: string;
  user_id: number;
  email: string;
  admin_tier: string;
  usd_to_satoshi_rate: null | number;
  web_intercom_hash: string;
  web_api_version: string;
  active_organisation_id: number;
  organisation_ids: string[];
}

export interface OrganisationLoginData extends TokenData {
  organisations: Organisation[];
  organisation: Organisation;
}

export interface AdminData {
  admin: AdminUser[];
  admins: AdminUser[];
  invites: Invite[];
  invite: Invite;
}